glib: Change API for repeating an in-thread scheduled task
Previously, in glib.schedule(), when in_thread is set to True, it is expected
that the task method return whether it wants to be repeated (True) or
not (False). However, this was not documented and different from the behavior of
the separate-thread task (passed as argument repeat=). To simply and improve
consistency, use the repeat= argument instead of return value from the method.

This fixes issue with daily diagnostics not running for the second time.

Signed-off-by: Sunil Mohan Adapa <sunil@medhas.org>